Adding Scatch project
The project was rushed and not what I would have liked to produce. Very eager to proceed with the course. The biggest bug I'm aware of is Barry hitting a wall and getting stuck in a 'correction' loop that I created to solve hitting walls. Also, not very efficient or clean I feel like(especially Barrys move block)